+++ ci/start_weaviate.sh
@@ -1,10 +1,7 @@
 #!/bin/bash
 
-echo "Download the Weaviate docker-compose file"
-curl -o docker-compose.yml "https://configuration.semi.technology/v2/docker-compose/docker-compose.yml?contextionary_compound_splitting=false&contextionary_language=en&contextionary_model=0.16.0&enterprise_usage_collector=false&media_type=text&qna_module=false&runtime=docker-compose&text_module=text2vec-contextionary&weaviate_version=v1.6.0"
-
 echo "Run Docker compose"
-nohup docker-compose up -d
+nohup docker-compose -f ci/docker-compose.yml up -d
 
 echo "Wait until weaviate is up"