+++ weaviate/gql/get.py
@@ -523,15 +523,63 @@ class GetBuilder(GraphQL):
         Examples
         --------
 
-        >>> # single additional property
+        >>> # single additional property with this GraphQL query
+        >>> '''
+        ... {
+        ...     Get {
+        ...         Article {
+        ...             title
+        ...             author
+        ...             _additional {
+        ...                 id
+        ...             }
+        ...         }
+        ...     }
+        ... }
+        ... '''
         >>> client.query\
         ...     .get('Article', ['title', 'author'])\
         ...     .with_additional('id']) # argument as `str`
-        >>> # multiple additional property
+
+        >>> # multiple additional property with this GraphQL query
+        >>> '''
+        ... {
+        ...     Get {
+        ...         Article {
+        ...             title
+        ...             author
+        ...             _additional {
+        ...                 id
+        ...                 certainty
+        ...             }
+        ...         }
+        ...     }
+        ... }
+        ... '''
         >>> client.query\
         ...     .get('Article', ['title', 'author'])\
         ...     .with_additional(['id', 'certainty']) # argument as `List[str]`
-        >>> # additional properties as clause
+
+        >>> # additional properties as clause with this GraphQL query
+        >>> '''
+        ... {
+        ...     Get {
+        ...         Article {
+        ...             title
+        ...             author
+        ...             _additional {
+        ...                 classification {
+        ...                     basedOn
+        ...                     classifiedFields
+        ...                     completed
+        ...                     id
+        ...                     scope
+        ...                 }
+        ...             }
+        ...         }
+        ...     }
+        ... }
+        ... '''
         >>> client.query\
         ...     .get('Article', ['title', 'author'])\
         ...     .with_additional(
@@ -539,32 +587,55 @@ class GetBuilder(GraphQL):
         ...             'classification' : ['basedOn', 'classifiedFields', 'completed', 'id']
         ...         }
         ...     ) # argument as `dict[str, List[str]]`
-        >>> # or
+        >>> # or with this GraphQL query
+        >>> '''
+        ... {
+        ...     Get {
+        ...         Article {
+        ...             title
+        ...             author
+        ...             _additional {
+        ...                 classification {
+        ...                     completed
+        ...                 }
+        ...             }
+        ...         }
+        ...     }
+        ... }
+        ... '''
         >>> client.query\
         ...     .get('Article', ['title', 'author'])\
         ...     .with_additional(
         ...         {
         ...             'classification' : 'completed'
         ...         }
-        ...     ) # argument as `dict[str, str]`
+        ...     ) # argument as `Dict[str, str]`
 
         Consider the following GraphQL clause:
-        >>> {
-        ...     _additional {
-        ...         token(
-        ...             properties: ["content"], # is required
-        ...             limit: 10, # optional, int
-        ...             certainty: 0.8 # optional, float
-        ...         ) {
-        ...             certainty
-        ...             endPosition
-        ...             entity
-        ...             property
-        ...             startPosition
-        ...             word
+        >>> '''
+        ... {
+        ...     Get {
+        ...         Article {
+        ...             title
+        ...             author
+        ...             _additional {
+        ...                 token (
+        ...                     properties: ["content"]
+        ...                     limit: 10
+        ...                     certainty: 0.8
+        ...                 ) {
+        ...                     certainty
+        ...                     endPosition
+        ...                     entity
+        ...                     property
+        ...                     startPosition
+        ...                     word
+        ...                 }
+        ...             }
         ...         }
         ...     }
         ... }
+        ... '''
 
         Then the python translation of this is the following:
         >>> clause = {
@@ -578,9 +649,9 @@ class GetBuilder(GraphQL):
         ...     ]
         ... }
         >>> settings = {
-        ...     'properties': ["content"], # is required
-        ...     'limit': 10, # optional, int
-        ...     'certainty': 0.8 # optional, float
+        ...     'properties': ["content"],  # is required
+        ...     'limit': 10,                # optional, int
+        ...     'certainty': 0.8            # optional, float
         ... }
         >>> client.query\
         ...     .get('Article', ['title', 'author'])\
@@ -588,7 +659,8 @@ class GetBuilder(GraphQL):
         ...         (clause, settings)
         ...     ) # argument as `Tuple[Dict[str, List[str]], Dict[str, Any]]`
 
-
+        If the desired clause does not match any example above, then the clause can always be
+        converted to string before passing it to the `.with_additional()` method.
 
         Returns
         -------