+++ docs/changelog.rst
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+Version 3.2.1
+-------------
+This patch adds support for ```array``` data types: `int[]`, `number[]`, `text[]`, `string[]`.
+
 Version 3.2.0
 -------------